pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Generic streaming RAG chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=1.5.0",
  "cryptography>=46.0.5",
  "fastapi>=0.129.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "langchain>=1.2.10",
  "langchain-anthropic>=1.3.3",
  "langchain-chroma>=1.1.0",
  "langchain-core>=1.2.12",
  "langchain-google-genai>=4.2.0",
  "langchain-groq>=1.1.2",
  "langchain-ollama>=1.0.1",
  "langchain-openai>=1.1.9",
  "langchain-text-splitters>=1.1.0",
  "psycopg[binary]>=3.3.2",
  "pydantic>=2.12.5",
  "pypdf>=6.7.0",
  "python-docx>=1.2.0",
  "python-dotenv>=1.2.1",
  "python-multipart>=0.0.22",
  "sqlalchemy>=2.0.46",
  "uvicorn[standard]>=0.40.0",
]
